+++ mwcp/__init__.py
@@ -23,4 +23,4 @@ from mwcp.core import run
 from mwcp.exceptions import *
 
 
-__version__ = "3.3.0"
+__version__ = "3.3.1"